medium-flask-app/pyproject.toml
[tool.poetry]
name = "my-name-app"
version = "0.1.0"
description = "dependencies for my-name-app"
authors = ["mishrasunny-coder <mishra.sunny@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
